instagramFinderFinal/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pytest==7.4.3
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
hikerapi
pytest-asyncio 